chore: improve Icon exports
